+++ app/Http/Controllers/Admin/CategoryController.php
@@ -16,6 +16,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        checkpermission('category');
         $data = Category::latest()->get();
         return view('admin.category.index',compact('data'));
     }
@@ -121,6 +122,7 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        checkpermission('category');
         Category::find($id)->delete();
         $notification=array(
             'messege'=>'Category Successfully Deleted !',
